Refactor UI helpers
